refactor(api): move legacy api wrapper into protocol_api/legacy
This puts it alongside where we'll implement the actual functionality.
