+++ api/src/opentrons/protocol_api/execute.py
@@ -6,7 +6,7 @@ from typing import Any, Callable
 
 import opentrons
 from .contexts import ProtocolContext
-from . import execute_v3, back_compat
+from . import execute_v3, legacy_wrapper
 
 from opentrons import config
 from opentrons.protocols.types import PythonProtocol, Protocol
@@ -121,7 +121,7 @@ def _run_python(
 def _run_python_legacy(proto: PythonProtocol, context: ProtocolContext):
     new_locs = locals()
     new_globs = globals()
-    namespace_mapping = back_compat.build_globals()
+    namespace_mapping = legacy_wrapper.api.build_globals()
     for key, value in namespace_mapping.items():
         setattr(opentrons, key, value)
     try:
